Refactor Semantic and ByteCodeOptimizer functions
The changes primarily constitute corrections in the ByteCodeOptimizer and enhancement in Semantic functions' parameters across the codebase. The changes in ByteCodeOptimizer involve altering some incorrect type usages while keeping the logic intact. In the Semantic module, the parameters in methods were better annotated for readability, and a particular condition's comparison value was nothing to UINT32_MAX.