Remove unnecessary .c_str() calls
The .c_str() calls were removed from the code, simplifying it and making clear that the variables are of Utf8 type, which already allows direct comparisons and other string operations. The change affects a wide range of files and functions. Replacing them with direct usage of the Utf8 variables provides a more straightforward and expressive notation.