Replace token id equality checks with new token methods
Changes were made in how tokens are compared with their IDs through the Parser and Semantic files. Instead of directly comparing the token id with a certain TokeId, two new methods are introduced within the TokenParse and Token classes - "is" and "isNot". These methods perform the TokenId comparison and improve code readability.