Add collectAutoScope function to Semantic class
The code has been refactored to add a new method, collectAutoScope, to the Semantic class. This method helps to decide the scoping depending on the context by scanning the parent hierarchy for an already defined type. The findIdentifierInScopes method was also updated to include a call to the new collectAutoScope method.