Update logging and documentation output methods
Modified code in various files to replace 'print' function with 'write' for logging and added a new command line option to control the output of generated documentation. The change in function naming is aimed to be more precise about the functionality of the code. Command line configuration for documentation output enables more control over the program's behavior.