Update compiler diagnostic flags and fix type inconsistencies
This change updates the compiler diagnostic flags in the pch.h file to ignore warnings related to switch statements and function with variable arguments. It also corrects a type inconsistency by casting the loop variable in ByteCodeRun_FFI.cpp to uint32_t and properly initializing the va_list in ErrorIds.cpp.