Add functionality to format and output function syntax trees
The changes include adding formatAST_Func file which contains functions specific to formatting the syntax trees of functions. Additionally, the FormatAst_Node file is updated to contain new nodes related to compiler languages. Further, the new functionality is integrated with the existing SyntaxJob and Parser. Finally, checks have been added to ensure that top-level generated namespaces are not redundantly processed or displayed.