Refactor function parameters and perform type promotion. REMOVE generic error with untyped (promote instead)
The function parameters for several functions in the Generic namespace have been reordered for greater clarity and consistency. In the Generic_Deduce.cpp file, the parameter order was changed for the 'deduceType' function and types are now promoted for 'typeInfoReplace' in GenericReplaceType. An added boolean argument was also placed at the end of the parameter list for 'replaceGenericParameters'. While performing this refactor, some test assertions were commented out in compiler4215.swg and compiler5427.swg for testing or debugging purposes.