Refactor handling of format flags in AstNode
Moved format flag handling from Parser to AstNode. A new function, inheritFormatFlags, is added to AstNode, which encapsulates the logic previously present in the Parser. This change improves the encapsulation and readability of the code. The other parts of the codebase that were using the old fmtFlags variable have been updated to use the new formatFlags function.