Add new file and refactor comment beautification
A new file, FormatAst_Beautify.cpp, was added to the project to isolate beautification operations from the rest of FormatAst functionalities. In consequence, beautifyComment function was refactored, replacing the pair structure with a new structure - "TokenComment" - to improve code readability. This change has been reflected in the project files accordingly.