Add functionality to format files and refactor 'scriptName' to 'fileName'
Implemented a new feature to format and beautify files. This includes a new command 'format' and function 'formatCommand'. Also, refactored all instances of the variable name 'scriptName' to 'fileName' to make it more general. An additional flag for 'Fake' modules was added, and the ability to determine if a source file accepts internal stuff was also introduced.