Improve handling of comment formatting and inheritance
Added functionality to handle comment and format inheritance in syntax sugar for multi-affectation. This involves changes in the Parser class and Tokenizer to adjust the usage of comments. As part of this process, the 'TokenComments' structure has been renamed to 'TokenFormat', reflecting the broader usage of this data structure. Additional functions were added to the AstNode class to inherit formatting elements from other nodes.