Update error handling in Parser and refine FormatJob
The Parser has been improved to handle errors more explicitly. Previously, it asserted that the current file was a CFG or script file. Now, it returns an error diagnostic if this is not the case. In addition, the FormatJob’s fileName type has been changed from Utf8 to Path, and variables in the execute() function have been appropriately renamed. Also added condition to handle script files with '.swgs' extension in temporary file load stage.