Update formatting of function call parameters
The format of function call parameters within the FormatAst_Literal.cpp code was updated. An if-else statement was added to differentiate the formatting between regular nodes and nodes that are parents of a FuncCallParam-type. The indenting and EOL characteristics are preserved for non-parent nodes, but are omitted for parents of FuncCallParam, thereby improving readability.