Refactor indent adding in FormatAst.cpp
Removed the general indent addition from the FormatAst.cpp code. The indent is now added only when the node does not have a SPEC_FLAG_PRIVATE. Additionally, replaced the 'outputNode' function call with the 'outputChildren' function when a node has a SPEC_FLAG_NO_CURLY, simplifying the code block for better readability.