Rename 'isForceTakeAddress' to 'setForceTakeAddress'
The function 'isForceTakeAddress' has been renamed to 'setForceTakeAddress' for clarity, as the former name could be misinterpreted as a boolean check. The method calls throughout the parser have been updated accordingly. Furthermore, a minor edit has been made to outputClosureArguments in FormatAst to utilize outputNode instead of directly adding string.