Add AST_GENERATED flag to freeze format in Ast.cpp
The code has been modified to generate an AST_GENERATED flag, specifically when the format is not frozen in Ast.cpp. This flag has been used within the Parser_Function.cpp to control the generation of new nodes and to manage the capturing of parameters. The patch intends to improve code flexibility and maintainability.