Add 'skipfmt' directive to skip source file formatting
The update introduces a new 'skipfmt' directive in the language. When this directive is encountered in a source file, it prevents the file from being formatted. The 'skipfmt' directive has been accommodated in Parser_Compiler, LanguageSpec and corresponding syntax files. The 'FILE_NO_FORMAT' flag in SourceFile is set when 'skipfmt' is encountered, causing the format job for the file to be cancelled.