Add const and var spec flags to AstVarDecl
This commit introduces new spec flags, SPEC_FLAG_FORCE_VAR and SPEC_FLAG_FORCE_CONST, to the AstVarDecl class. This allows for finer control over variable declaration nodes in the Parser. It also modifies the FormatAst_Type module to properly format these new spec flag changes in AstVarDecl. This results in the correct representation of const and var specifier in the output.