Adjust indentation in Parser_Struct and limit AST_GENERATED flag addition
The code has been adjusted to improve readability by realigning some elements in the Parser_Struct. Also, a new condition has been introduced to add the AST_GENERATED flag only to a structNode if the file is not intended for formatting. This precaution prevents potential misuse of the AST_GENERATED flag.