Update format of non-private namespace block
Code block format has been updated for non-private namespaces. The logic has been revised so that indentation is only added if the node does not have the AstNameSpace::SPEC_FLAG_PRIVATE flag. The flag check for AstNameSpace::SPEC_FLAG_USING remains the same, but it's nested within the non-private condition to maintain the intended print order.