Update test, refactor imports, and add returnType check
The code has been updated to un-skip a test and change its assertion method in the compiler5430 test. The imports in ByteCodeGen_Function.cpp have been reordered for better organization. Also, a check was added in Semantic_Var.cpp to adjust the function return type if the function return type is not yet defined.