Add 'ignore-bad-params' flag and support for extra arguments
The commit includes the addition of 'ignore-bad-params' flag to 'scripts.bat', 'web.bat', 'scripts_dm.bat', and 'web_dm.bat' calls in the 'x64.bat' and 'x64_dm.bat' files. Also, it extends the number of arguments that can be passed to 'swag doc' and 'swag_devmode doc' in 'web.bat' and 'web_dm.bat' files.