Update code to use pointer instead of const
Mutability has been added to the nodes within the code. In place of using constant nodes, the code has been refactored to make use of mutable pointers. This update primarily impacts the FormatAst and GenDoc modules where references to AstNode have been switched from `const AstNode*` to `AstNode*`. This allows for more flexibility and dynamic functioning of the nodes within the codebase.