Improve code formatting and flag handling
The changes made in multiple files enhance the code formatting. A special focus is placed on upgrading the way enum values are outputted and flagged which now includes aligning them based on their length. Updates were made in the AstNode and FormatAst modules to handle flags using .add() method instead of direct assignment. Furthermore, improvements were made on how blanks and indents are added in the FormatConcat module.