Refactor parsing and formatting methods
In this commit, multiple Methods like `outputChildrenEnumValues`, `outputChildrenVar`, in the `FormatAst` class and the `FormatConcat::clear` method have been updated for clarity and to avoid redundancies. Additionally, an assertion is now checked in the `Semantic_Type.cpp` file to ensure the existence of the `resolvedSymbolOverload` method. This is to streamline the format parsing and improve the efficiency of the code.