Refactor FormatOptions to FormatContext in various files
Replaced the existing FormatOptions structure with FormatContext and adjusted its usage throughout the code. This change refactors the formatting settings into a context that is laid out in a way that is easier to manage and more streamlined in handling formatting options. Moved some settings such as 'outputComments' and 'outputBlankLines' that were previously in FormatOptions to FormatContext to consolidate options.