Fix alignment and comment handling in code
The commit includes a fix to the alignment of 'szInfoTitle' in shell32.swg and uuid.swg files. Additionally, a condition is added in FormatAst_Var.cpp to ensure comments are aligned only when the type didn't output some eol. This prevents faulty code formatting which could impact the code readability.