Add format tracking in Tokenizer and Parser
The `trimMultilineString` function has been moved into the Tokenizer class and modified to not execute when format tracking is turned on. The Tokenizer and Parser classes have been updated to include a new flag for format tracking. This allows the accurate alignment of multi-line strings during processing, as well as better handling of end of line characters.