Refactor code for checking end-of-line in nodes
This commit introduces a new function `hasEOLInside` in `FormatAst` that checks whether an `AstNode` has an end-of-line inside it. This function is used in `FormatAst_Struct` to simplify the code. The `Parser_Var` class also has some changes to remove unneeded assignment. Modifications are also made to `FormatAst_Attr` to handle `NULL` child nodes properly.