Add solution for Advent of Code 2016 day 10 problem
The new code added, serves as a solution for the "Balance Bots" problem from Advent of Code 2016, day 10. The implementation involves creating a structure to represent a bot and functions to manipulate the bot's actions such as adding values, eating low or high values. Also, the logic to solve the puzzle using the bot actions is implemented.