Refactor code structure in 10AB.swg
This refactor specifies the 'Bot' struct with additional 'pending' Array field and refactors the 'day10AB' function for cleaner code structure. The function 'execute' has been created to handle Array of strings and added to improve code efficiency. Also, 'bots' var moved to wider scope to be accesible for 'getBot' and 'execute' methods.