Refactor regex objects and update variable setting in 7B.swg
In the aoc2017 module, the code in 7B.swg has been streamlined for better readability and performance. The redundant regex object (`r`) was renamed for better context (to `r0`) and the variable setting method was updated for `set.add` to directly include the extraction and conversion of values, eliminating the need for an extra variable. A minor arrangement of code lines was done for improved logical flow.