Update documentation and add new Array methods
Several documentation updates in html files were carried out. This included an adjustment of the generation date, as well as the addition of new Array methods such as "fill" and "growResize" in the methods list on the std.core.html page. Additionally, detailed descriptions for these new methods were added.