Refactor symbol searching in scope hierarchy
Updated the method in which scope hierarchies symbols are searched - loops are now placed for each symbol lookup to ensure full coverage. The code orders the procedures to restart until all types are completed. This revision also includes the handling of cases regarding incomplete types during collection of scope hierarchy.