Add day10A solution to Advent of Code 2017 module
This commit includes the solution for day10A of the Advent of Code 2017 challenge. The solution involves the implementation of a hash function simulation based on knot-tying. A file containing the necessary input data is also included in the commit. In addition, a check is added to the reverse function to skip reversal if a slice has one or zero elements.