Refactor code and add `reduce` function to slice treatment
The code added the `reduce` function to the slice treatment module, which allows for the transformation of the content of a slice with a given lambda. This change helps improve code readability and efficiency. Moreover, there are enhancements in handling operator assignment for strings, runes, and u8 arrays in the string module. Another functionality changed is in 10B.swg module, where numbers are now added as a list instead of individually to the 'nums' variable.