Advent of Code 2019 Day 10 A Add solution for monitoring station location problem
This commit introduces code for solving the problem of selecting the optimal location for a monitoring station within an asteroid field. The solution involves creating a map of asteroid locations, calculating potential sight lines, and determining the location that offers visibility to the most asteroids. The main logic is implemented in the day10A() function. Additionally, this commit includes max and min functions to the slice module, and changes the module year in the aoc.bat script.