Add asteroid vaporization feature and update math functions
Implemented a new function in the aoc2019 module for the asteroid vaporization feature in the 'day10B' file. Additionally, the math functions in 'vector2.swg' have been updated to improve the calculation of vector angles. The main function in the 'main.swg' file was also updated to include the newly-created function.