Add day 10 challenge solution for 2020 AoC
This commit introduces the solution for day 10 challenge of the Advent of Code 2020. It includes a new file containing the problem input (10.txt) and the solution script (10A.swg). Additionally, the main execution script (main.swg) and the batch execution file (aoc.bat) have been updated to include the new challenge.