Add day 10, part B solution for Advent of Code 2020
A new file has been added for the Advent of Code 2020 day 10 part B problem solution. The function `day10B()` is defined in the 10B.swg file under the aoc2020 module. Also, the method call for `day10B()` has been added to the main.swg file to ensure it runs as part of the solution sequence.