Add color picker functionality to gui module
A new color picker functionality has been added to the gui module. This includes changes to the `propertcolor.swg` file, introducing various functions that are fundamental for creating and handling a color picker. The `color.swg` file within the pixel module was also updated for this purpose, making the `isOpaque()` method constant.