Add support for enhanced monitor and device capabilities
Additional constants representing device capabilities have been included to the gdi32 module. A function GetDeviceCaps has also been added to the module. The Monitor structure in the core module has been updated to account for scaling and to include a real monitor and work rectangle. This structure is then utilized in the win32 core module to provide a more accurate representation of the physical monitor.